Fix LogEntry parsing for Nuki 4
Nuki 3 uses padding in LogEntry, while Nuki 4 doesn't

Simply making the padding optional doesn't as Optional() cannot work in
internal Struct if the external Struct has fields after this Struct (as
Optional() is greedy).

To handle this, we limit the "payload" field in NukiMessage Struct using
OffsettedEnd() and seek back to the end of actually parsed data so we can
catch all unhandled data.